move print settings menu to a new file
